Adicionando parametro para definição de tipo de retorno e nova regra de negocio para criação de usuarios

+++ app/backend/api-produto/src/main/java/com/conture/apiproduto/repository/CategoriaProdutoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CategoriaProdutoRepository extends JpaRepository<CategoriaProduto, Long> {
+public interface CategoriaProdutoRepository extends JpaRepository<CategoriaProduto, Integer> {
 	Optional<CategoriaProduto> findByNomeIgnoreCase(String nome);
 }
