+++ app/backend/api-produto/src/main/java/com/conture/apiproduto/repository/PreferenciaDonatarioRepository.java
@@ -3,5 +3,5 @@ package com.conture.apiproduto.repository;
 import com.conture.apiproduto.entity.PreferenciaDonatario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PreferenciaDonatarioRepository extends JpaRepository<PreferenciaDonatario, Long> {
+public interface PreferenciaDonatarioRepository extends JpaRepository<PreferenciaDonatario, Integer> {
 }
